Extract StateSearch.
